requirements.txt
Pillow
pillow_avif_plugin
pillow_heif
PyQt6
PyQt6_sip
